🚚 api: [PRESCRIPTION] move service to prescription/campaign context
