+++ api/src/prescription/campaign/domain/read-models/CampaignAnalysis.js
@@ -1,6 +1,6 @@
 import _ from 'lodash';
 
-import * as recommendationService from '../../../../../lib/domain/services/recommendation-service.js';
+import * as recommendationService from '../services/recommendation-service.js';
 
 class CampaignAnalysis {
   constructor({ campaignId, campaignLearningContent, tutorials, participantCount = 0 } = {}) {
